# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DigitPro Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transfers:
    get:
      operationId: getMyTransfers
      tags: [transfers]
      summary: Get user's transfer simulations
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"

    post:
      operationId: createTransfer
      tags: [transfers]
      summary: Create a transfer simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferRequest"
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}:
    get:
      operationId: getTransfer
      tags: [transfers]
      summary: Get a transfer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/pdf:
    get:
      operationId: getTransferPdf
      tags: [transfers]
      summary: Download transfer PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /transfers/track/{trackingCode}:
    get:
      operationId: trackTransfer
      tags: [transfers]
      summary: Track a transfer by tracking code (public)
      parameters:
        - name: trackingCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferTracking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    CreateTransferRequest:
      type: object
      properties:
        type:
          type: string
          enum: [simulation, document]
        senderName:
          type: string
        senderIban:
          type: string
        senderBank:
          type: string
        receiverName:
          type: string
        receiverIban:
          type: string
        receiverBank:
          type: string
        amount:
          type: number
        currency:
          type: string
        reference:
          type: string
        description:
          type: string
        transferDate:
          type: string
          format: date
        estimatedArrival:
          type: string
          format: date
      required:
        - type
        - senderName
        - senderBank
        - receiverName
        - receiverBank
        - amount
        - currency
    Transfer:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [simulation, document]
        trackingCode:
          type: string
        trackingUrl:
          type: string
        senderName:
          type: string
        senderIban:
          type: string
        senderBank:
          type: string
        receiverName:
          type: string
        receiverIban:
          type: string
        receiverBank:
          type: string
        amount:
          type: number
        currency:
          type: string
        reference:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [initiated, processing, completed, failed]
        progress:
          type: integer
        transferDate:
          type: string
          format: date
        estimatedArrival:
          type: string
          format: date
        coinCost:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - trackingCode
        - trackingUrl
        - senderName
        - senderBank
        - receiverName
        - receiverBank
        - amount
        - currency
        - status
        - progress
        - coinCost
        - createdAt
    TransferTracking:
      type: object
      properties:
        trackingCode:
          type: string
        senderName:
          type: string
        senderBank:
          type: string
        receiverName:
          type: string
        receiverBank:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [initiated, processing, completed, failed]
        progress:
          type: integer
        transferDate:
          type: string
          format: date
        estimatedArrival:
          type: string
          format: date
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TransferStep"
      required:
        - trackingCode
        - senderName
        - senderBank
        - receiverName
        - receiverBank
        - amount
        - currency
        - status
        - progress
        - steps
    TransferStep:
      type: object
      properties:
        label:
          type: string
        description:
          type: string
        completed:
          type: boolean
        timestamp:
          type: string
          format: date-time
      required:
        - label
        - description
        - completed