# lib/api-zod/src/generated/types/createTransferRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigitPro Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransferRequestType } from "./createTransferRequestType";

export interface CreateTransferRequest {
  type: CreateTransferRequestType;
  senderName: string;
  senderIban?: string;
  senderBank: string;
  receiverName: string;
  receiverIban?: string;
  receiverBank: string;
  amount: number;
  currency: string;
  reference?: string;
  description?: string;
  transferDate?: Date;
  estimatedArrival?: Date;
}