# lib/api-zod/src/generated/types/transferStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigitPro Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TransferStep {
  label: string;
  description: string;
  completed: boolean;
  timestamp?: Date;
}